requirements.txt
PyYAML
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
google-generativeai
Flask
markdown2
